refactor: move implicit-LOD validation to ShaderMixer.Validation.cs
Keep the validation standalone, in its own partial file, next to the
other ShaderMixer partials. Pure move, no behaviour change.
